Add logger and fix zip file bug

+++ gw/install/EmeraldAccInstaller.py
@@ -6,6 +6,12 @@ import zipfile
 import codecs
 import re
 import urllib.request
+from logging import getLogger,StreamHandler,DEBUG,INFO
+logger = getLogger(__name__)
+handler = StreamHandler()
+handler.setLevel(DEBUG)
+logger.setLevel(INFO)
+logger.addHandler(handler)
 
 parser = argparse.ArgumentParser(description='Unzip accelerator zip file to targetDir. Merge *.properties file. TargetDir must be root dir contains modules folder.')
 parser.add_argument('-i', '--input', required=True, help=r'zip file path or build url. e.g.) -i c:\tmp\base-module.zip or -i http://thfiles/builds/PC/8.0/8.0.3.270_20150527.0303_618017/')
@@ -17,9 +23,13 @@ print(args)
 
 # Make zip url from build url http://thfiles/builds/PC/8.0/8.0.3.270_20150527.0303_618017/
 def makeZipURLFromBuild(url, module):
-    return  url + 'jrdc/' + module + '-module.zip'
+    jrdc = '/jrdc/'
+    if url.endswith('/'):
+        jrdc = 'jrdc/'
+    return  url + jrdc + module + '-module.zip'
 
 def downloadFromWeb(url):
+    logger.info('download file: '+ url)
     local_filename, headers =urllib.request.urlretrieve(url)
     return local_filename
 
@@ -35,7 +45,7 @@ def readFileToMap(file, dic=None):
         if len(arr) != 2:
             continue
         if arr[0] in properties and properties[arr[0]].rstrip() != arr[1].rstrip():
-            print('Duplicate key %s, old=%s new=%s' % (arr[0], properties[arr[0]], arr[1]))
+            logger.info('Duplicate key %s, old=%s new=%s' % (arr[0], properties[arr[0]], arr[1]))
         properties[arr[0]] = arr[1]
     return properties
 
@@ -51,7 +61,7 @@ def readLinesToMap(chunk, dic=None):
         if len(arr) != 2:
             continue
         if arr[0] in properties and properties[arr[0]].rstrip() != arr[1].rstrip():
-            print('Duplicate key %s, old=%s new=%s' % (arr[0], properties[arr[0]], arr[1]))
+            logger.info('Duplicate key %s, old=%s new=%s' % (arr[0], properties[arr[0]], arr[1]))
         properties[arr[0]] = arr[1]
     return properties
 
@@ -63,7 +73,7 @@ def unzip(zip_filename):
     zip_file = zipfile.ZipFile(zip_filename, "r")
     for filename in zip_file.namelist():
         if not os.path.basename(filename):
-            print('makedir '+ filename)
+            logger.debug('makedir '+ filename)
             path = os.path.join(target,filename)
             if not os.path.exists(path):
                 os.mkdir(path)
@@ -85,15 +95,17 @@ def unzip(zip_filename):
 
 # main
 configuration = 'modules/configuration'
-modules = args.modules.split(',')
 target = os.path.join(args.out, configuration)
 if args.input.startswith('http'):
+    if args.modules is None:
+        exit('modules parameter is required for url')
+    modules = args.modules.split(',')
     for module in modules:
         #make each url then download zip
         zipPath = downloadFromWeb(makeZipURLFromBuild(args.input, module))
-        print('Temp file save at ' + zipPath)
+        logger.info('Temp file save at ' + zipPath)
         unzip(zipPath)
-        print('Remove temp file : '+ zipPath)
+        logger.info('Remove temp file : '+ zipPath)
         os.remove(zipPath)
 else:
-    unzip(args.zipFile)
+    unzip(args.input)
